Better value decoding
This commit changes how values are decoded internally. It introduces the
`ValueDecoder` trait, which is responsible for decoding a value from either
text or binary using a particular strategy.

Additionally, OIDs are now used to map to the name of the postgres receive
function for a type, which is a reliable way of deciding how the value should
be decoded. These receive functions are then mapped to ValueDecoders.  This
change allows for cusom value decoders to be specified, and also allows for
the remote database to be queried for its types; custom types may be present
with a non-fixed OID, but which use existing receive functions.  Thus, type
handling is more robust against postgres use cases.

Changes to the user-facing API:
- Client.apply now accepts additional optional arguments:
  - `customReceiveFunctions` which allows for specifying a `PartialFunction` that
    maps receive function names to `ValueDecoder`s
  - `binaryResults` which allows specifying that the client should ask the server
    for results in binary format (which is more robust and performant than text)
- Timestamps now use appropriate `java.time` representations by default:
  - `java.time.LocalDateTime` for timestamp
  - `java.time.ZonedDateTime` for timestamp with time zone
  - `java.time.LocalDate` for date

Compatibility changes:
- Java 8 is now required (since `java.time` is used)

The integration tests have also been changed to use an environment-configurable
Postgres instance, rather than a hard-coded one.  This will allow the integration
tests to actually run in travis-ci.

Tests also now use ScalaCheck where possible.

There are also a few fixes for bugs which didn't become apparent until binary
mode was possible.