Fix PreparedStatement to set a unique name
PostgreSQL prepared statements and portals are referenced by names, where
setting an empty string is ok, as it means it is unnamed prepared statement. [1]

However there is a weird issue happening under high loads where a different set
of parameters are being set to the wrong prepared statements:

`bind message supplies 10 parameters, but prepared statement “...” requires 12`

And also sporadically we get `ERROR: portal "" cannot be run` making it super
hard to troubleshoot as all portals currently share the same name.

Setting a unique name for each prepared statement has fixed the problem in our
production setting, however I am not able to pinpoint why the problem exists in
the first place, as "" seems to be a valid portal name.

[1] - https://www.postgresql.org/docs/11/protocol-flow.html#PROTOCOL-FLOW-EXT-QUERY