Support sending SET commands.
Previous to this change, if you tried to issue a `SET` command,
finagle-postgres would raise an exception telling no completion command
is known for tag `SET`. This patch implements such a response from
server.

`SET` commands are needed for sending runtime settings to postgres
or when trying to set the current transaction isolation level.

See:

https://github.com/getquill/quill/issues/29#issuecomment-560162692